fix : int -> size_t

+++ srcs/service/event_handler.cc
@@ -44,7 +44,7 @@ void Service::RunCommand(int const epoll_fd, CommandMapping const &cmd_map,
       (*(cmd_map.find(command)->second))(cur_event.data.fd, msg);
 
   // set all messages
-  for (int i = 0; i < fd_to_write.size(); ++i)
+  for (size_t i = 0; i < fd_to_write.size(); ++i)
     SetWriteEvent(epoll_fd, fd_to_write[i]);
 }
 
@@ -71,7 +71,7 @@ void Service::HandleClientEvent(int const epoll_fd,
 
     // run commands
     if (messages.has_value()) {
-      for (int i = 0; i < (*messages).size(); ++i) {
+      for (size_t i = 0; i < (*messages).size(); ++i) {
         RunCommand(epoll_fd, cmd_map, cur_event, (*messages)[i]);
       }
     }
